Move cell implementation to separate class
